requirements.txt
async-timeout==5.0.1
psycopg2==2.9.10
python-dotenv==1.0.1
redis==5.2.0
